move redis and db to drivers directory
