+++ pkg/api/auth/service.go
@@ -13,7 +13,7 @@ import (
 
 	"github.com/andreas-kokkalis/dock_server/pkg/api"
 	"github.com/andreas-kokkalis/dock_server/pkg/api/store"
-	"github.com/andreas-kokkalis/dock_server/pkg/db"
+	"github.com/andreas-kokkalis/dock_server/pkg/drivers/db"
 	"github.com/jordic/lti"
 	"github.com/julienschmidt/httprouter"
 )
