+++ pkg/api/store/redis_repo.go
@@ -11,7 +11,7 @@ import (
 	"time"
 
 	"github.com/andreas-kokkalis/dock_server/pkg/api"
-	"github.com/andreas-kokkalis/dock_server/pkg/cache"
+	"github.com/andreas-kokkalis/dock_server/pkg/drivers/cache"
 )
 
 // RedisRepo ...
