+++ pkg/integration/spec_test.go
@@ -5,8 +5,8 @@ import (
 	"testing"
 
 	"github.com/andreas-kokkalis/dock_server/cmd/dock_server/schema/dbutil"
-	"github.com/andreas-kokkalis/dock_server/pkg/cache/redismock"
-	"github.com/andreas-kokkalis/dock_server/pkg/db"
+	"github.com/andreas-kokkalis/dock_server/pkg/drivers/cache/redismock"
+	"github.com/andreas-kokkalis/dock_server/pkg/drivers/db"
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
 	"github.com/stretchr/testify/assert"
